halley-wl: split reusable window helpers into decoration and geometry modules
- move decoration metrics and border-rect construction into window/decoration.rs
- move geometry snapshots, surface sync helpers, and render-path crop/dst utilities into window/geometry.rs
- update window/mod.rs to consume the new modules
- keep collect_active_surfaces in window/mod.rs
- preserve capture.rs and stack.rs shared-import flow
